+++ includes/admin/class-rop-rest-api.php
@@ -1153,7 +1153,7 @@ class Rop_Rest_Api {
 	private function add_account_buffer( $data ) {
 		$services        = array();
 		$active_accounts = array();
-		$buffer_service = new Rop_Buffer_Service();
+		$buffer_service  = new Rop_Buffer_Service();
 		$model           = new Rop_Services_Model();
 		$db              = new Rop_Db_Upgrade();
 
@@ -1186,6 +1186,103 @@ class Rop_Rest_Api {
 		} else {
 			update_option( $rop_buffer_via_rs_app_option, 'true' );
 		}
+
+		return $this->response->to_array();
+	}
+
+	/**
+	 * API method called to add Tumblr Blogs via app.
+	 *
+	 * @SuppressWarnings(PHPMD.UnusedPrivateMethod) As it is called dynamically.
+	 *
+	 * @since   8.5.7
+	 * @access  private
+	 *
+	 * @param   array $data Tumblr accounts data.
+	 *
+	 * @return  array
+	 */
+	private function add_account_tumblr( $data ) {
+		$services        = array();
+		$active_accounts = array();
+		$tumblr_service  = new Rop_Tumblr_Service();
+		$model           = new Rop_Services_Model();
+		$db              = new Rop_Db_Upgrade();
+
+		$tumblr_service->add_account_with_app( $data );
+
+		$services[ $tumblr_service->get_service_id() ] = $tumblr_service->get_service();
+		$active_accounts                                = array_merge( $active_accounts, $tumblr_service->get_service_active_accounts() );
+
+		if ( ! empty( $services ) ) {
+			$model->add_authenticated_service( $services );
+		}
+
+		if ( ! empty( $active_accounts ) ) {
+			$db->migrate_schedule( $active_accounts );
+			$db->migrate_post_formats( $active_accounts );
+		} else {
+			$this->response->set_code( '500' )
+						   ->set_data( array() );
+
+			return $this->response->to_array();
+		}
+
+		$this->response->set_code( '200' )
+					   ->set_message( 'OK' )
+					   ->set_data( array() );
+
+		$rop_tumblr_via_rs_app_option = 'rop_tumblr_via_rs_app';
+		if ( ! get_option( $rop_tumblr_via_rs_app_option ) ) {
+			add_option( $rop_tumblr_via_rs_app_option, 'true', ' ', 'no' );
+		} else {
+			update_option( $rop_tumblr_via_rs_app_option, 'true' );
+		}
+
+		return $this->response->to_array();
+	}
+
+	/**
+	 * API method called to add Google My Business locations via app.
+	 *
+	 * @SuppressWarnings(PHPMD.UnusedPrivateMethod) As it is called dynamically.
+	 *
+	 * @since   8.5.9
+	 * @access  private
+	 *
+	 * @param   array $data Tumblr accounts data.
+	 *
+	 * @return  array
+	 */
+	private function add_account_gmb( $data ) {
+		$services        = array();
+		$active_accounts = array();
+		$gmb_service     = new Rop_Gmb_Service();
+		$model           = new Rop_Services_Model();
+		$db              = new Rop_Db_Upgrade();
+
+		$gmb_service->add_account_with_app( $data );
+
+		$services[ $gmb_service->get_service_id() ] = $gmb_service->get_service();
+		$active_accounts                                = array_merge( $active_accounts, $gmb_service->get_service_active_accounts() );
+
+		if ( ! empty( $services ) ) {
+			$model->add_authenticated_service( $services );
+		}
+
+		if ( ! empty( $active_accounts ) ) {
+			$db->migrate_schedule( $active_accounts );
+			$db->migrate_post_formats( $active_accounts );
+		} else {
+			$this->response->set_code( '500' )
+						   ->set_data( array() );
+
+			return $this->response->to_array();
+		}
+
+		$this->response->set_code( '200' )
+					   ->set_message( 'OK' )
+					   ->set_data( array() );
 
 		return $this->response->to_array();
 	}