+++ includes/admin/helpers/class-rop-cron-helper.php
@@ -41,7 +41,6 @@ class Rop_Cron_Helper {
 	 * @return mixed
 	 * @since   8.0.0
 	 * @access  public
-	 *
 	 */
 	public static function rop_cron_schedules( $schedules ) {
 		$schedules['5min'] = array(
@@ -149,7 +148,6 @@ class Rop_Cron_Helper {
 	 * @return bool
 	 * @since   8.0.0rc
 	 * @access  public
-	 *
 	 */
 	public function create_cron( $first = true ) {
 		if ( defined( 'ROP_CRON_ALTERNATIVE' ) && false === ROP_CRON_ALTERNATIVE && ! wp_next_scheduled( self::CRON_NAMESPACE ) ) {
@@ -160,6 +158,12 @@ class Rop_Cron_Helper {
 				wp_schedule_single_event( time() + 30, self::CRON_NAMESPACE_ONCE );
 			}
 			wp_schedule_event( time(), '5min', self::CRON_NAMESPACE );
+			/**
+			 * Changing this option to true, upon page refresh the WP Cron Jobs will work as normal.
+			 * This value must become true anytime the "Start Share" button is clicked.
+			 *
+			 * @see Rop_Admin::check_cron_status()
+			 */
 			$this->cron_status_global_change( true );
 
 		} elseif ( defined( 'ROP_CRON_ALTERNATIVE' ) && true === ROP_CRON_ALTERNATIVE ) {
@@ -168,6 +172,12 @@ class Rop_Cron_Helper {
 				$this->fresh_start();
 				$settings = new Rop_Global_Settings();
 				$settings->update_start_time();
+				/**
+				 * Changing this option to true, upon page refresh the WP Cron Jobs will work as normal.
+				 * This value must become true anytime the "Start Share" button is clicked.
+				 *
+				 * @see Rop_Admin::check_cron_status()
+				 */
 				$this->cron_status_global_change( true );
 			}
 		}
@@ -183,19 +193,23 @@ class Rop_Cron_Helper {
 	 * @return bool
 	 * @since   8.0.0rc
 	 * @access  public
-	 *
 	 */
 	public function remove_cron( $request = array() ) {
 		global $wpdb;
-		// Mark sharing system as stopped
 
+		/**
+		 * Changing this option to false, upon page refresh the WP Cron Jobs will be cleared.
+		 * This value must become false anytime the "Stop Share" button is clicked.
+		 *
+		 * @see Rop_Admin::check_cron_status()
+		 */
 		$this->cron_status_global_change( false );
 
 		$current_cron_list = _get_cron_array();
 		$rop_cron_key      = self::get_schedule_key( array( self::CRON_NAMESPACE, self::CRON_NAMESPACE_ONCE ) );
+
 		if ( ! empty( $rop_cron_key ) ) {
 			$wpdb->query( 'START TRANSACTION' );
-
 			foreach ( $rop_cron_key as $rop_active_cron ) {
 				$cron_time      = (int) $rop_active_cron['time'];
 				$cron_key       = $rop_active_cron['key'];
@@ -233,7 +247,6 @@ class Rop_Cron_Helper {
 	 *
 	 * @see wp_unschedule_event()
 	 * @see _get_cron_array()
-	 *
 	 */
 	public static function get_schedule_key( $namespace ) {
 		if ( empty( $namespace ) ) {
@@ -278,7 +291,6 @@ class Rop_Cron_Helper {
 	 * @param bool $action true/false if crons should work or stop.
 	 *
 	 * @since 8.5.0
-	 *
 	 */
 	function cron_status_global_change( $action = false ) {
 		$key         = 'rop_is_sharing_cron_active';