+++ vue/src/vue-elements/sign-in-btn.vue
@@ -7,8 +7,9 @@
 					class="btn input-group-btn"
 					:class="'btn-' + network"
 					@click="requestAuthorization( network )">
-				<i v-if="network !== 'buffer'" class="fa fa-fw" :class="'fa-' + network"></i>
+				<i v-if="network !== 'buffer' && network !== 'gmb'" class="fa fa-fw" :class="'fa-' + network"></i>
 				<i v-if="network === 'buffer'" class="fa fa-fw fa-plus-square"></i>
+				<i v-if="network === 'gmb'" class="fa fa-fw fa-google"></i>
 				{{service.name}}
 			</button>
 
@@ -25,23 +26,30 @@
 					<div class="content">
 						<div class="auth-app" v-if="isFacebook">
 							<button class="btn btn-primary big-btn" @click="openPopupFB()">{{labels.fb_app_signin_btn}}</button>
-							<span class="text-center">{{labels.fb_own_app_signin}}</span>
+							<span class="text-center">{{labels.app_option_signin}}</span>
 						</div>
 						<div class="auth-app" v-if="isTwitter">
 							<button class="btn btn-primary big-btn" @click="openPopupTW()">{{labels.tw_app_signin_btn}}</button>
-							<span class="text-center">{{labels.tw_own_app_signin}}</span>
+							<span class="text-center">{{labels.app_option_signin}}</span>
 						</div>
-						<div class="auth-app" v-if="isLinkedIn && isAllowedLinkedIn">
+						<div class="auth-app" v-if="isLinkedIn">
 							<button class="btn btn-primary big-btn" @click="openPopupLI()">{{labels.li_app_signin_btn}}</button>
-							<span class="text-center">{{labels.li_own_app_signin}}</span>
+							<span class="text-center">{{labels.app_option_signin}}</span>
+						</div>
+						<div class="auth-app" v-if="isTumblr && isAllowedTumblr">
+							<button class="btn btn-primary big-btn" @click="openPopupTumblr()">{{labels.tumblr_app_signin_btn}}</button>
+							<span class="text-center">{{labels.app_option_signin}}</span>
 						</div>
 						<div class="auth-app" v-if="isBuffer">
 							<button class="btn btn-primary big-btn" @click="openPopupBuffer()">{{labels.buffer_app_signin_btn}}</button>
 						</div>
-						<div id="rop-advanced-config" v-if="isFacebook || isTwitter || (isLinkedIn && isAllowedLinkedIn)">
+						<div class="auth-app" v-if="isGmb">
+							<button class="btn btn-primary big-btn" id="gmb-btn" @click="openPopupGmb()">{{labels.gmb_app_signin_btn}}</button>
+						</div>
+						<div id="rop-advanced-config" v-if="isFacebook || isTwitter || isLinkedIn || (isTumblr && isAllowedTumblr)">
 						<button class="btn btn-primary" v-on:click="showAdvanceConfig = !showAdvanceConfig">{{labels.show_advance_config}}</button>
 					</div>
-						<div v-if="showAdvanceConfig && (isFacebook || isTwitter || (isLinkedIn && isAllowedLinkedIn) )">
+						<div v-if="showAdvanceConfig && (isFacebook || isTwitter || isLinkedIn || (isTumblr && isAllowedTumblr) )">
 						<div class="form-group" v-for="( field, id ) in modal.data">
 							<label class="form-label" :for="field.id">{{ field.name }}</label>
 							<input :class="[ 'form-input', field.error ? ' is-error' : '' ]" type="text" :id="field.id" v-model="field.value"
@@ -50,7 +58,7 @@
 							<p class="text-gray">{{ field.description }}</p>
 						</div>
 					</div>
-						<div v-if="(!isTwitter && !isFacebook && !isLinkedIn && !isBuffer) || (isLinkedIn && !isAllowedLinkedIn)">
+						<div v-if="(!isTwitter && !isFacebook && !isLinkedIn && !isBuffer && !isGmb && !isTumblr) || (isTumblr && !isAllowedTumblr)">
 						<div class="form-group" v-for="( field, id ) in modal.data">
 							<label class="form-label" :for="field.id">{{ field.name }}</label>
 							<input :class="[ 'form-input', field.error ? ' is-error' : '' ]" type="text" :id="field.id" v-model="field.value"
@@ -61,14 +69,14 @@
 					</div>
 				</div>
 				</div>
-				<div v-if="isFacebook || isTwitter || (isLinkedIn && isAllowedLinkedIn)" class="modal-footer">
+				<div v-if="isFacebook || isTwitter || isLinkedIn || isBuffer || isGmb || (isTumblr && isAllowedTumblr)" class="modal-footer">
 					<p class="text-left pull-left mr-2" v-html="labels.rs_app_info"></p>
 				</div>
-				<div v-if="showAdvanceConfig && (isFacebook || isTwitter || isLinkedIn)" class="modal-footer">
+				<div v-if="showAdvanceConfig && (isFacebook || isTwitter || isLinkedIn || isTumblr)" class="modal-footer">
 					<div class="text-left pull-left mr-2" v-html="modal.description"></div>
 					<button class="btn btn-primary" @click="closeModal()">{{labels.sign_in_btn}}</button>
 				</div>
-				<div v-if="(!isTwitter && !isFacebook && !isLinkedIn && !isBuffer) || (isLinkedIn && !isAllowedLinkedIn)" class="modal-footer">
+				<div v-if="(!isTwitter && !isFacebook && !isLinkedIn && !isBuffer && !isGmb && !isTumblr) || (isTumblr && !isAllowedTumblr)" class="modal-footer">
 					<div class="text-left pull-left mr-2" v-html="modal.description"></div>
 					<button class="btn btn-primary" @click="closeModal()">{{labels.sign_in_btn}}</button>
 				</div>
@@ -99,6 +107,8 @@
         appPathTW: ropAuthAppData.authAppTwitterPath,
         appPathLI: ropAuthAppData.authAppLinkedInPath,
         appPathBuffer: ropAuthAppData.authAppBufferPath,
+        appPathTumblr: ropAuthAppData.authAppTumblrPath,
+        appPathGmb: ropAuthAppData.authAppGmbPath,
 				appAdminEmail: ropAuthAppData.adminEmail,
 				siteAdminUrl: ropAuthAppData.adminUrl,
 				appUniqueId: ropAuthAppData.authToken,
@@ -106,6 +116,7 @@
 				windowParameters: 'top=20,left=100,width=560,height=670',
 				authPopupWindow: null,
 				showLiAppBtn: ropApiSettings.show_li_app_btn,
+				showTmblrAppBtn: ropApiSettings.show_tmblr_app_btn,
 				showBtn: false
 			}
 		},
@@ -333,6 +344,44 @@
                     Vue.$log.error('Got nothing from server. Prompt user to check internet connection and try again', error)
 				});
 			},
+            /**
+             * Add Tumblr account.
+             *
+             * @param data Data.
+             */
+            addAccountTumblr(data) {
+                this.$store.dispatch('fetchAJAXPromise', {
+                    req: 'add_account_tumblr',
+                    updateState: false,
+                    data: data
+                }).then(response => {
+                    window.removeEventListener("message", event => this.getChildWindowMessage(event));
+                    this.authPopupWindow.close();
+                    window.location.reload();
+                }, error => {
+                    this.is_loading = false;
+                    Vue.$log.error('Got nothing from server. Prompt user to check internet connection and try again', error)
+				});
+			},
+            /**
+             * Add Google My Business account.
+             *
+             * @param data Data.
+             */
+            addAccountGmb(data) {
+                this.$store.dispatch('fetchAJAXPromise', {
+                    req: 'add_account_gmb',
+                    updateState: false,
+                    data: data
+                }).then(response => {
+                    window.removeEventListener("message", event => this.getChildWindowMessage(event));
+                    this.authPopupWindow.close();
+                    window.location.reload();
+                }, error => {
+                    this.is_loading = false;
+                    Vue.$log.error('Got nothing from server. Prompt user to check internet connection and try again', error)
+				});
+			},
 			getChildWindowMessage: function (event) {
 				if (~event.origin.indexOf(this.appOrigin)) {
             if ('Twitter' === this.modal.serviceName) {
@@ -343,6 +392,10 @@
 					    this.addAccountLI(JSON.parse(event.data));
 						} else if ('Buffer' === this.modal.serviceName) {
 					    this.addAccountBuffer(JSON.parse(event.data));
+            } else if ('Tumblr' === this.modal.serviceName) {
+					    this.addAccountTumblr(JSON.parse(event.data));
+            } else if ('Gmb' === this.modal.serviceName) {
+					    this.addAccountGmb(JSON.parse(event.data));
             }
 
 				} else {
@@ -396,7 +449,31 @@
                     this.cancelModal();
                 }
                 window.addEventListener("message", event => this.getChildWindowMessage(event));
-			}
+			},
+            openPopupTumblr: function () { // Open the popup specific for Tumblr
+                let loginUrl = this.appOrigin + this.appPathTumblr + '?callback_url=' + this.siteAdminUrl + '&token=' + this.appUniqueId + '&signature=' + this.appSignature + '&data=' + this.appAdminEmail;
+                try {
+                    this.authPopupWindow.close();
+                } catch (e) {
+                    // nothing to do
+                } finally {
+                    this.authPopupWindow = window.open(loginUrl, 'authTmblr', this.windowParameters);
+                    this.cancelModal();
+                }
+                window.addEventListener("message", event => this.getChildWindowMessage(event));
+			},
+            openPopupGmb: function () { // Open the popup specific for Tumblr
+                let loginUrl = this.appOrigin + this.appPathGmb + '?callback_url=' + this.siteAdminUrl + '&token=' + this.appUniqueId + '&signature=' + this.appSignature + '&data=' + this.appAdminEmail;
+                try {
+                    this.authPopupWindow.close();
+                } catch (e) {
+                    // nothing to do
+                } finally {
+                    this.authPopupWindow = window.open(loginUrl, 'authGmb', this.windowParameters);
+                    this.cancelModal();
+                }
+                window.addEventListener("message", event => this.getChildWindowMessage(event));
+			},
 		},
 		computed: {
 			selected_service: function () {
@@ -436,17 +513,30 @@
             isLinkedIn() {
                 return this.modal.serviceName === 'LinkedIn';
             },
-						isAllowedLinkedIn: function () {
+						// will return true if the current service actions are for Buffer.
+						isBuffer() {
+								return this.modal.serviceName === 'Buffer';
+						},
+            // will return true if the current service actions are for Tumblr.
+            isTumblr() {
+                return this.modal.serviceName === 'Tumblr';
+            },
+            // will return true if the current service actions are for Google My Business.
+            isGmb() {
+                return this.modal.serviceName === 'Gmb';
+            },
+            // will return true if the current service actions are for Pinterest.
+            isPinterest() {
+                return this.modal.serviceName === 'Pinterest';
+            },
+						isAllowedTumblr: function () {
 							let showButton = true;
-							if (!this.showLiAppBtn) {
+							if (!this.showTmblrAppBtn) {
 									showButton = false;
 							}
 							return showButton;
 					},
-					// will return true if the current service actions are for Buffer.
-					isBuffer() {
-							return this.modal.serviceName === 'Buffer';
-					},
+
 	}
 }
 </script>
@@ -455,5 +545,11 @@
 		cursor:not-allowed;
 		pointer-events: auto;
 		opacity: 0.3;
+	}
+	.big-btn#gmb-btn{
+	padding: 0 35px 0 14px;
+	}
+	.btn-gmb{
+	text-transform: uppercase;
 	}
 </style>